Voltage level issue
This image shows how to raise the voltage level of 3V digital output to 5V. Instead, in this project, please ensure that the source S pin of the 2N7000 is NOT grounded but instead connected to the 3V power source, while the end of the 10k resistor not connected to the transistor is connected to 3V ground. This ensures that the final output voltage is not inverted. 3V source used to ensure raise to only 5V and not more. The Gate G pin of the same transistor is connected to the digital output pin of the nodemcu which will be used to trigger the required analog sensor. Finally, the point between Drain D pin and the 10 k resistor, which is the 5V output, is connected to the sensor to activate/turn on. Also, when making connections, ensure that the Vcc pins of the non-analog sensors are connected to 5V source, and given 5V ground, while the analog sensors (Gas and WaterLevel) are given the digital output pin triggers (as shown in arduino IDE code) to their Vcc pins, and given 3V source ground to their ground pins.